+++ lib/apexcharts/charts/range_bar.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module ApexCharts
+  class RangeBarChart < BarChart
+    def chart_type
+      'rangeBar'
+    end
+  end
+end